Map uuid -> id
